// api/generated/model/getRecipesDifficulty.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Cheffy API
 * OpenAPI spec version: 1.0.0
 */

export type GetRecipesDifficulty = typeof GetRecipesDifficulty[keyof typeof GetRecipesDifficulty];


export const GetRecipesDifficulty = {
  EASY: 'EASY',
  MEDIUM: 'MEDIUM',
  HARD: 'HARD',
  EXPERT: 'EXPERT',
} as const;